Current working status for map -> graph processing

+++ z2p/map_data.py
@@ -31,7 +31,7 @@ class RomMap:
 
         self._sub_map_data = []
         self.map_data = np.zeros(
-            [2 * self.SUB_MAP_SIZE_X, 2 * self.SUB_MAP_SIZE_Y], dtype=np.int
+            [2 * self.SUB_MAP_SIZE_X, 2 * self.SUB_MAP_SIZE_Y], dtype=int
         )
         if os.path.isfile(rompath) and os.path.exists(rompath):
             self.romdata = os.path.abspath(rompath)
